backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
rq==1.11.0
redis==4.6.0
boto3==1.28.32
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
sqlalchemy==1.4.55
psycopg2-binary==2.9.7
pydantic==1.10.11
python-multipart==0.0.6
